fix(manager): switch manager tests to use singleDC cluster
Since there is an issue with multiDC cluster restore when the EaR is
turned on (https://github.com/scylladb/scylla-manager/issues/3829),
it was decided to temporarily switch the main part of jobs to run on
singleDC cluster. Only one multiDC cluster job is left for enterprise
version 2022 where EaR is not implemented.