feature(vector-search): add tests for Vector Search in Cloud
The change introduces a few tests for Vector Search verification in
Scylla Cloud. The tests are based on old siren-tests + SCT integration
and expect Cloud cluster to be prepared in advance by siren-tests.

New tests include:
- sanity checks for VS (add some Vector data into cluster, build index
and run some ANN search request for verification)
- replace node scenarios to make sure VS node can be replaced without
any lose of availability for healthy node and with some loses for broken
node

sct_config.py extended with two new parameters which comes from
siren-tests and can be retrieved in SCT currently:
- vector_search_node_ids - IDs of deployed VS nodes
- vector_search_node_type_id - ID of deployed VS nodes

In addition to tests, a new email reporter added to serve VS tests.